ARGO_MDP/requirements.txt
# ARGO MDP Requirements

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
gym>=0.21.0

# Optional: For LLM integration
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
